app/requirements.txt
Flask==3.0.0
pandas==2.1.4
Pillow==10.2.0
uvicorn[standard]==0.22.0
